move resolve logic in existing resolver module
